+++ changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - New method `delete_event` in customer-api v3.6.
 - New methods in configuration-api v3.6 for greetings: `create_greeting`, `delete_greeting`, `get_greeting`, `update_greeting`, `list_greetings`.
+- New method `send_thinking_indicator` in agent-api v3.6.
 
 ### Changed
 - Improved websocket response collection + extended logging in the websocket client.